refactor(ui): streamline MainWindow workflows and optimize tree filtering
