+++ Apps/Avalonia/DevProjex.Avalonia/MainWindow.axaml.cs
@@ -113,7 +113,6 @@ public partial class MainWindow : Window
     private readonly FilterOptionSelectionService _filterSelectionService;
     private readonly TreeExportService _treeExport;
     private readonly SelectedContentExportService _contentExport;
-    private readonly TreeAndContentExportService _treeAndContentExport;
     private readonly ProjectCopyExportService _projectCopyExport;
     private readonly PreviewDocumentBuilder _previewDocumentBuilder;
     private readonly RepositoryWebPathPresentationService _repositoryWebPathPresentationService;
@@ -143,6 +142,7 @@ public partial class MainWindow : Window
     private readonly ProjectLoadSnapshotPipeline _projectLoadSnapshotPipeline;
     private readonly PreviewWorkspacePipeline _previewPipeline;
     private readonly RefreshTreePipeline _refreshPipeline;
+    private readonly ProjectTextOutputPipeline _textOutputPipeline;
     private readonly ProjectProfilePersistenceCoordinator _projectProfiles;
     private readonly ProjectLoadCancellationCoordinator _projectLoadCancellation = new();
     private readonly TaskbarProgressCoordinator _taskbarProgress;
@@ -150,14 +150,7 @@ public partial class MainWindow : Window
     private BuildTreeResult? _currentTree;
     private BuildTreeResult? _filterBaseTree;
     private ProjectTreeInventoryState? _currentTreeInventory;
-    private TreeNodeDescriptor? _lastInteractiveFilteredRoot;
-    private TreeNodeDescriptor? _lastInteractiveFilterBaseRoot;
-    private string? _lastInteractiveFilterQuery;
     private bool _lastInteractiveFilterUsedInMemory;
-    private readonly Dictionary<string, TreeNodeDescriptor> _interactiveFilterQueryCache = new(StringComparer.OrdinalIgnoreCase);
-    private readonly LinkedList<string> _interactiveFilterQueryCacheLru = [];
-    private readonly Dictionary<string, LinkedListNode<string>> _interactiveFilterQueryCacheNodes = new(StringComparer.OrdinalIgnoreCase);
-    private const int InteractiveFilterQueryCacheLimit = 8;
     // Advanced ignore counts are always part of the ignore-options UX now. The old
     // persisted toggle is normalized to true so legacy settings cannot hide counts.
     private const bool AdvancedIgnoreCountsAlwaysEnabled = true;
@@ -380,7 +373,6 @@ public partial class MainWindow : Window
         _filterSelectionService = services.FilterOptionSelectionService;
         _treeExport = services.TreeExportService;
         _contentExport = services.ContentExportService;
-        _treeAndContentExport = services.TreeAndContentExportService;
         _projectCopyExport = services.ProjectCopyExportService;
         _previewDocumentBuilder = services.PreviewDocumentBuilder;
         _repositoryWebPathPresentationService = services.RepositoryWebPathPresentationService;
@@ -424,6 +416,10 @@ public partial class MainWindow : Window
             // 350ms delay ensures thumb animation (250ms) completes fully before loading.
             UiTimingProfile.Scale(TimeSpan.FromMilliseconds(350)));
         _refreshPipeline = new RefreshTreePipeline(this);
+        _textOutputPipeline = new ProjectTextOutputPipeline(
+            _treeExport,
+            _contentExport,
+            services.TreeAndContentExportService);
         _selectionCoordinator = new SelectionSyncCoordinator(
             _viewModel,
             _scanOptions,
@@ -2560,332 +2556,6 @@ public partial class MainWindow : Window
 
     private void OnExit(object? sender, RoutedEventArgs e) => Close();
 
-    private async void OnCopyTree(object? sender, RoutedEventArgs e)
-    {
-        try
-        {
-            if (!EnsureTreeReady()) return;
-
-            var selected = GetCheckedPaths();
-            var format = GetCurrentTreeTextFormat();
-            var content = BuildTreeTextForSelection(selected, format);
-
-            await SetClipboardTextAsync(content);
-            _sessionMetrics.RecordClipboard("tree", format, content.Length, success: true);
-            _toastService.Show(_localization["Toast.Copy.Tree"]);
-        }
-        catch (Exception ex)
-        {
-            await ShowErrorAsync(ex.Message);
-        }
-    }
-
-    private async void OnCopyContent(object? sender, RoutedEventArgs e)
-    {
-        long? statusOperationId = null;
-        try
-        {
-            if (!EnsureTreeReady()) return;
-
-            // Cancel background metrics calculation - user wants immediate action
-            _metrics.CancelBackgroundCalculation();
-
-            var selected = GetCheckedPaths();
-            var files = BuildOrderedUniqueFilePaths(selected);
-
-            if (files.Count == 0)
-            {
-                if (selected.Count > 0)
-                    await ShowInfoAsync(_localization["Msg.NoCheckedFiles"]);
-                else
-                    await ShowInfoAsync(_localization["Msg.NoTextContent"]);
-                return;
-            }
-
-            // Run file reading off UI thread
-            statusOperationId = BeginOutputPreparationStatus();
-            var pathPresentation = CreateExportPathPresentation();
-            var content = await Task.Run(() => _contentExport.BuildAsync(
-                files,
-                CancellationToken.None,
-                pathPresentation?.MapFilePath));
-            if (string.IsNullOrWhiteSpace(content))
-            {
-                CompleteStatusOperation(ref statusOperationId);
-                await ShowInfoAsync(_localization["Msg.NoTextContent"]);
-                return;
-            }
-
-            await SetClipboardTextAsync(content);
-            _sessionMetrics.RecordClipboard("content", format: null, content.Length, success: true);
-            CompleteStatusOperation(ref statusOperationId);
-            _toastService.Show(_localization["Toast.Copy.Content"]);
-        }
-        catch (Exception ex)
-        {
-            CompleteStatusOperation(ref statusOperationId);
-            await ShowErrorAsync(ex.Message);
-        }
-    }
-
-    private async void OnCopyTreeAndContent(object? sender, RoutedEventArgs e)
-    {
-        long? statusOperationId = null;
-        try
-        {
-            if (!EnsureTreeReady()) return;
-
-            // Cancel background metrics calculation - user wants immediate action
-            _metrics.CancelBackgroundCalculation();
-
-            var selected = GetCheckedPaths();
-            var format = GetCurrentTreeTextFormat();
-            var pathPresentation = CreateExportPathPresentation();
-            // Run file reading off UI thread
-            statusOperationId = BeginOutputPreparationStatus();
-            var content = await Task.Run(() =>
-                _treeAndContentExport.BuildAsync(
-                    _currentPath!,
-                    _currentTree!.Root,
-                    selected,
-                    format,
-                    CancellationToken.None,
-                    pathPresentation));
-            await SetClipboardTextAsync(content);
-            _sessionMetrics.RecordClipboard("tree-content", format, content.Length, success: true);
-            CompleteStatusOperation(ref statusOperationId);
-            _toastService.Show(_localization["Toast.Copy.TreeAndContent"]);
-        }
-        catch (Exception ex)
-        {
-            CompleteStatusOperation(ref statusOperationId);
-            await ShowErrorAsync(ex.Message);
-        }
-    }
-
-    private async void OnExportTreeToFile(object? sender, RoutedEventArgs e)
-    {
-        try
-        {
-            if (!EnsureTreeReady()) return;
-
-            var selected = GetCheckedPaths();
-            var format = GetCurrentTreeTextFormat();
-            var content = BuildTreeTextForSelection(selected, format);
-
-            var saved = await TryExportTextToFileAsync(
-                content,
-                BuildSuggestedExportFileName("tree", GetTreeExportFileExtension(format)),
-                _viewModel.MenuFileExportTree,
-                defaultExtension: GetTreeExportFileExtension(format),
-                fileTypeChoices: CreateTreeExportFileTypeChoices(format));
-
-            if (saved)
-            {
-                _sessionMetrics.RecordFileExport("tree", format, content.Length, success: true);
-                _toastService.Show(_localization["Toast.Export.Tree"]);
-            }
-        }
-        catch (Exception ex)
-        {
-            await ShowErrorAsync(ex.Message);
-        }
-    }
-
-    private async void OnExportContentToFile(object? sender, RoutedEventArgs e)
-    {
-        long? statusOperationId = null;
-        try
-        {
-            if (!EnsureTreeReady()) return;
-
-            _metrics.CancelBackgroundCalculation();
-
-            var selected = GetCheckedPaths();
-            var files = BuildOrderedUniqueFilePaths(selected);
-
-            if (files.Count == 0)
-            {
-                if (selected.Count > 0)
-                    await ShowInfoAsync(_localization["Msg.NoCheckedFiles"]);
-                else
-                    await ShowInfoAsync(_localization["Msg.NoTextContent"]);
-                return;
-            }
-
-            statusOperationId = BeginOutputPreparationStatus();
-            var pathPresentation = CreateExportPathPresentation();
-            var content = await Task.Run(() => _contentExport.BuildAsync(
-                files,
-                CancellationToken.None,
-                pathPresentation?.MapFilePath));
-            if (string.IsNullOrWhiteSpace(content))
-            {
-                CompleteStatusOperation(ref statusOperationId);
-                await ShowInfoAsync(_localization["Msg.NoTextContent"]);
-                return;
-            }
-
-            // File preparation is complete. The native picker can remain open indefinitely
-            // while the user chooses a name, so it must never keep the app status busy.
-            CompleteStatusOperation(ref statusOperationId);
-            var saved = await TryExportTextToFileAsync(
-                content,
-                BuildSuggestedExportFileName("content", "txt"),
-                _viewModel.MenuFileExportContent,
-                defaultExtension: "txt",
-                fileTypeChoices: [CreateTextFileType()]);
-
-            if (saved)
-            {
-                _sessionMetrics.RecordFileExport("content", format: null, content.Length, success: true);
-                _toastService.Show(_localization["Toast.Export.Content"]);
-            }
-        }
-        catch (Exception ex)
-        {
-            CompleteStatusOperation(ref statusOperationId);
-            await ShowErrorAsync(ex.Message);
-        }
-    }
-
-    private async void OnExportTreeAndContentToFile(object? sender, RoutedEventArgs e)
-    {
-        long? statusOperationId = null;
-        try
-        {
-            if (!EnsureTreeReady()) return;
-
-            _metrics.CancelBackgroundCalculation();
-
-            var selected = GetCheckedPaths();
-            var format = GetCurrentTreeTextFormat();
-            var pathPresentation = CreateExportPathPresentation();
-
-            statusOperationId = BeginOutputPreparationStatus();
-            var content = await Task.Run(() =>
-                _treeAndContentExport.BuildAsync(
-                    _currentPath!,
-                    _currentTree!.Root,
-                    selected,
-                    format,
-                    CancellationToken.None,
-                    pathPresentation));
-
-            // A combined payload is plain text even when its leading tree is JSON/XML/MD.
-            // Finish the operation before opening the picker and keep the .txt contract.
-            CompleteStatusOperation(ref statusOperationId);
-            var saved = await TryExportTextToFileAsync(
-                content,
-                BuildSuggestedExportFileName("tree_content", "txt"),
-                _viewModel.MenuFileExportTreeAndContent,
-                defaultExtension: "txt",
-                fileTypeChoices: [CreateTextFileType()]);
-
-            if (saved)
-            {
-                _sessionMetrics.RecordFileExport("tree-content", format, content.Length, success: true);
-                _toastService.Show(_localization["Toast.Export.TreeAndContent"]);
-            }
-        }
-        catch (Exception ex)
-        {
-            CompleteStatusOperation(ref statusOperationId);
-            await ShowErrorAsync(ex.Message);
-        }
-    }
-
-    private TreeTextFormat GetCurrentTreeTextFormat()
-        => _viewModel.SelectedExportFormat switch
-        {
-            ExportFormat.Json => TreeTextFormat.Json,
-            ExportFormat.Xml => TreeTextFormat.Xml,
-            ExportFormat.Markdown => TreeTextFormat.Markdown,
-            _ => TreeTextFormat.Ascii
-        };
-
-    private string BuildTreeTextForSelection(IReadOnlySet<string> selectedPaths, TreeTextFormat format)
-    {
-        if (_currentTree is null || string.IsNullOrWhiteSpace(_currentPath))
-            return string.Empty;
-
-        var pathPresentation = CreateExportPathPresentation();
-        var displayRootPath = pathPresentation?.DisplayRootPath;
-        var displayRootName = pathPresentation?.DisplayRootName;
-        var hasSelection = selectedPaths.Count > 0;
-        var treeText = hasSelection
-            ? _treeExport.BuildSelectedTree(
-                _currentPath,
-                _currentTree.Root,
-                selectedPaths,
-                format,
-                displayRootPath,
-                displayRootName)
-            : _treeExport.BuildFullTree(
-                _currentPath,
-                _currentTree.Root,
-                format,
-                displayRootPath,
-                displayRootName);
-
-        if (hasSelection && string.IsNullOrWhiteSpace(treeText))
-            treeText = _treeExport.BuildFullTree(
-                _currentPath,
-                _currentTree.Root,
-                format,
-                displayRootPath,
-                displayRootName);
-
-        return treeText;
-    }
-
-    private ExportPathPresentation? CreateExportPathPresentation()
-    {
-        if (!_viewModel.IsGitMode)
-        {
-            _cachedPathPresentation = null;
-            _cachedPathPresentationProjectPath = null;
-            _cachedPathPresentationRepositoryUrl = null;
-            return null;
-        }
-
-        if (string.IsNullOrWhiteSpace(_currentPath) || string.IsNullOrWhiteSpace(_currentRepositoryUrl))
-        {
-            _cachedPathPresentation = null;
-            _cachedPathPresentationProjectPath = null;
-            _cachedPathPresentationRepositoryUrl = null;
-            return null;
-        }
-
-        if (_cachedPathPresentation is not null &&
-            string.Equals(_cachedPathPresentationProjectPath, _currentPath, StringComparison.Ordinal) &&
-            string.Equals(_cachedPathPresentationRepositoryUrl, _currentRepositoryUrl, StringComparison.Ordinal))
-        {
-            return _cachedPathPresentation;
-        }
-
-        _cachedPathPresentation = _repositoryWebPathPresentationService.TryCreate(_currentPath, _currentRepositoryUrl);
-        _cachedPathPresentationProjectPath = _currentPath;
-        _cachedPathPresentationRepositoryUrl = _currentRepositoryUrl;
-
-        return _cachedPathPresentation;
-    }
-
-    private static string MapExportDisplayPath(string filePath, Func<string, string>? mapFilePath)
-    {
-        if (mapFilePath is null)
-            return filePath;
-
-        try
-        {
-            var mapped = mapFilePath(filePath);
-            return string.IsNullOrWhiteSpace(mapped) ? filePath : mapped;
-        }
-        catch
-        {
-            return filePath;
-        }
-    }
 
     private void SchedulePreviewRefresh(bool immediate = false)
     {
@@ -3231,30 +2901,15 @@ public partial class MainWindow : Window
                 !string.IsNullOrWhiteSpace(currentPath) &&
                 currentTreeRoot is not null)
             {
-                var treeText = selectedPaths.Count > 0
-                    ? _treeExport.BuildSelectedTree(
+                var treeText = _textOutputPipeline.BuildTree(
+                    new ProjectTextOutputSnapshot(
                         currentPath,
                         currentTreeRoot,
                         selectedPaths,
+                        OrderedFilePaths: null,
                         treeFormat,
-                        pathPresentation?.DisplayRootPath,
-                        pathPresentation?.DisplayRootName)
-                    : _treeExport.BuildFullTree(
-                        currentPath,
-                        currentTreeRoot,
-                        treeFormat,
-                        pathPresentation?.DisplayRootPath,
-                        pathPresentation?.DisplayRootName);
-
-                if (selectedPaths.Count > 0 && string.IsNullOrWhiteSpace(treeText))
-                {
-                    treeText = _treeExport.BuildFullTree(
-                        currentPath,
-                        currentTreeRoot,
-                        treeFormat,
-                        pathPresentation?.DisplayRootPath,
-                        pathPresentation?.DisplayRootName);
-                }
+                        pathPresentation),
+                    cancellationToken);
 
                 if (string.IsNullOrWhiteSpace(treeText))
                     return null;
@@ -3390,7 +3045,17 @@ public partial class MainWindow : Window
 
         if (selectedMode == PreviewContentMode.Tree)
         {
-            var treePreviewText = BuildTreeTextForSelection(selectedPaths, treeFormat);
+            var treePreviewText = !string.IsNullOrWhiteSpace(currentPath) && currentTreeRoot is not null
+                ? _textOutputPipeline.BuildTree(
+                    new ProjectTextOutputSnapshot(
+                        currentPath,
+                        currentTreeRoot,
+                        selectedPaths,
+                        currentTreeOrderedFilePaths,
+                        treeFormat,
+                        pathPresentation),
+                    cancellationToken)
+                : string.Empty;
             var effectiveTreeText = string.IsNullOrEmpty(treePreviewText) ? noDataText : treePreviewText;
             return new PreviewBuildResult(_previewDocumentBuilder.CreateInMemory(effectiveTreeText));
         }
@@ -3511,144 +3176,6 @@ public partial class MainWindow : Window
     private static bool ShouldForcePreviewMemoryCleanup(long textLength, int lineCount) =>
         PreviewFileCollectionPolicy.ShouldForcePreviewMemoryCleanup(textLength, lineCount);
 
-    private async Task<bool> TryExportTextToFileAsync(
-        string content,
-        string suggestedFileName,
-        string dialogTitle,
-        string defaultExtension,
-        IReadOnlyList<FilePickerFileType> fileTypeChoices)
-    {
-        if (StorageProvider is null || string.IsNullOrWhiteSpace(content))
-            return false;
-
-        var options = new FilePickerSaveOptions
-        {
-            Title = dialogTitle,
-            SuggestedFileName = suggestedFileName,
-            ShowOverwritePrompt = true,
-            DefaultExtension = defaultExtension,
-            FileTypeChoices = fileTypeChoices
-        };
-
-        var file = await StorageProvider.SaveFilePickerAsync(options);
-        if (file is null)
-            return false;
-
-        var destinationPath = file.TryGetLocalPath();
-        if (!string.IsNullOrWhiteSpace(destinationPath) &&
-            !string.IsNullOrWhiteSpace(_currentPath))
-        {
-            try
-            {
-                // Text and physical exports share the same canonical guard so an aliased path
-                // cannot bypass the read-only boundary of the loaded project.
-                ProjectCopyExportService.EnsureDestinationOutsideProject(_currentPath, destinationPath);
-            }
-            catch (ProjectCopyExportException)
-            {
-                _toastService.Show(_localization["Error.ProjectCopy.UnsafeDestinationPath"]);
-                return false;
-            }
-        }
-
-        await using var stream = await file.OpenWriteAsync();
-        await _textFileExport.WriteAsync(stream, content);
-
-        return true;
-    }
-
-    private static IReadOnlyList<FilePickerFileType> CreateTreeExportFileTypeChoices(TreeTextFormat format)
-    {
-        if (format == TreeTextFormat.Ascii)
-            return [CreateTextFileType()];
-
-        var nativeFileType = format switch
-        {
-            TreeTextFormat.Json => CreateJsonFileType(),
-            TreeTextFormat.Xml => CreateXmlFileType(),
-            TreeTextFormat.Markdown => CreateMarkdownFileType(),
-            _ => CreateTextFileType()
-        };
-
-        // Structured tree text is also useful as a generic text artifact. Keep the native
-        // extension first while offering TXT as an explicit, semantically honest fallback.
-        return [nativeFileType, CreateTextFileType()];
-    }
-
-    private static FilePickerFileType CreateTextFileType()
-        => new("TXT")
-        {
-            Patterns = ["*.txt"],
-            MimeTypes = ["text/plain"]
-        };
-
-    private static FilePickerFileType CreateJsonFileType()
-        => new("JSON")
-        {
-            Patterns = ["*.json"],
-            MimeTypes = ["application/json"]
-        };
-
-    private static FilePickerFileType CreateXmlFileType()
-        => new("XML")
-        {
-            Patterns = ["*.xml"],
-            MimeTypes = ["application/xml", "text/xml"]
-        };
-
-    private static FilePickerFileType CreateMarkdownFileType()
-        => new("Markdown")
-        {
-            Patterns = ["*.md"],
-            MimeTypes = ["text/markdown", "text/plain"]
-        };
-
-    private static string GetTreeExportFileExtension(TreeTextFormat format)
-        => format switch
-        {
-            TreeTextFormat.Json => "json",
-            TreeTextFormat.Xml => "xml",
-            TreeTextFormat.Markdown => "md",
-            _ => "txt"
-        };
-
-    private void CompleteStatusOperation(ref long? operationId)
-    {
-        if (!operationId.HasValue)
-            return;
-
-        _statusOperations.Complete(operationId.Value);
-        operationId = null;
-    }
-
-    private long? BeginOutputPreparationStatus()
-    {
-        // Clipboard and text-file exports are safe against the captured project tree, but
-        // they must not replace the progress or cancellation action of a physical export.
-        if (_viewModel.IsProjectCopyExportInProgress)
-            return null;
-
-        return _statusOperations.Begin(
-            _localization["Status.Operation.PreparingOutput"],
-            indeterminate: true);
-    }
-
-    private string BuildSuggestedExportFileName(string suffix, string extension)
-    {
-        var baseName = _currentProjectDisplayName;
-        if (string.IsNullOrWhiteSpace(baseName) && !string.IsNullOrWhiteSpace(_currentPath))
-            baseName = Path.GetFileName(_currentPath.TrimEnd(Path.DirectorySeparatorChar, Path.AltDirectorySeparatorChar));
-
-        if (string.IsNullOrWhiteSpace(baseName))
-            baseName = "devprojex";
-
-        var invalidChars = Path.GetInvalidFileNameChars();
-        var sanitized = new StringBuilder(baseName.Length);
-        foreach (var ch in baseName)
-            sanitized.Append(invalidChars.Contains(ch) ? '_' : ch);
-
-        return $"{sanitized}_{suffix}.{extension}";
-    }
 
     private void OnExpandAll(object? sender, RoutedEventArgs e)
     {
@@ -5605,801 +5132,7 @@ public partial class MainWindow : Window
         _themeBrushCoordinator.UpdateDynamicThemeBrushes();
     }
 
-    #region Recent Projects
 
-    private void LoadRecentProjects()
-    {
-        _recentProjectsDb = _recentProjectsStore.LoadForStartup(StartupStoreLockTimeout);
-        SyncRecentProjectsToViewModel();
-    }
-
-    private void SyncRecentProjectsToViewModel()
-    {
-        _viewModel.RecentFolders.Clear();
-        foreach (var entry in _recentProjectsDb.RecentFolders)
-        {
-            _viewModel.RecentFolders.Add(new RecentProjectEntryViewModel(
-                entry.Path,
-                RecentProjectPresentationService.CreateFolderDisplayText(entry.Path),
-                RecentProjectPresentationService.CreateFolderToolTip(entry.Path)));
-        }
-
-        _viewModel.RecentRepositories.Clear();
-        foreach (var entry in _recentProjectsDb.RecentRepositories)
-        {
-            _viewModel.RecentRepositories.Add(new RecentProjectEntryViewModel(
-                entry.Url,
-                RecentProjectPresentationService.CreateRepositoryDisplayText(entry.Url),
-                RecentProjectPresentationService.CreateRepositoryToolTip(entry.Url)));
-        }
-    }
-
-    private void AttachRecentMenuHandlers()
-    {
-        if (_topMenuBar?.RecentMenuItemControl is { } recentMenuItem)
-            recentMenuItem.SubmenuOpened += OnRecentMenuSubmenuOpened;
-    }
-
-    private void DetachRecentMenuHandlers()
-    {
-        if (_topMenuBar?.RecentMenuItemControl is { } recentMenuItem)
-            recentMenuItem.SubmenuOpened -= OnRecentMenuSubmenuOpened;
-    }
-
-    private void OnRecentMenuSubmenuOpened(object? sender, RoutedEventArgs e)
-    {
-        RefreshRecentFoldersMenu();
-        StartRecentFolderAvailabilityRefresh();
-    }
-
-    private void RefreshRecentFoldersMenu()
-    {
-        var recentMenuItem = _topMenuBar?.RecentMenuItemControl;
-        if (recentMenuItem is null)
-            return;
-
-        recentMenuItem.Items.Clear();
-
-        if (_viewModel.RecentFolders.Count == 0)
-        {
-            recentMenuItem.Items.Add(new MenuItem
-            {
-                Header = _viewModel.MenuFileRecentEmpty,
-                IsEnabled = false
-            });
-            return;
-        }
-
-        foreach (var recentFolder in _viewModel.RecentFolders)
-        {
-            var item = new MenuItem
-            {
-                Header = recentFolder.DisplayText,
-                Tag = recentFolder.Value
-            };
-
-            ToolTip.SetTip(item, null);
-            SetRecentFolderMenuItemAvailability(
-                item,
-                !_unavailableRecentFolderPaths.Contains(recentFolder.Value));
-            item.Click += OnRecentFolderMenuItemClick;
-            recentMenuItem.Items.Add(item);
-        }
-    }
-
-    private async void OnRecentFolderMenuItemClick(object? sender, RoutedEventArgs e)
-    {
-        if (!_viewModel.CanChangeProjectTree || sender is not MenuItem { Tag: string path })
-            return;
-
-        var lifetimeToken = _windowLifetimeCts?.Token ?? CancellationToken.None;
-        bool isAvailable;
-        try
-        {
-            isAvailable = await _recentFolderAvailabilityService.IsAvailableAsync(path, lifetimeToken);
-        }
-        catch (OperationCanceledException)
-        {
-            return;
-        }
-
-        UpdateRecentFolderAvailability(path, isAvailable);
-        ApplyRecentFolderAvailabilityToMenu();
-        if (!isAvailable)
-        {
-            var shouldRemove = await MessageDialog.ShowConfirmationAsync(
-                this,
-                _localization["Dialog.RecentFolderUnavailable.Title"],
-                _localization.Format("Dialog.RecentFolderUnavailable.Message", path),
-                _localization["Dialog.RecentFolderUnavailable.Remove"],
-                _localization["Dialog.RecentFolderUnavailable.Keep"],
-                width: 450,
-                height: 180);
-
-            if (shouldRemove)
-                RemoveRecentFolder(path);
-            return;
-        }
-
-        await TryOpenFolderAsync(path, fromDialog: true);
-    }
-
-    private void StartRecentFolderAvailabilityRefresh()
-    {
-        if (_recentFolderAvailabilityRefreshTask is { IsCompleted: false } ||
-            _windowLifetimeCts is not { } lifetime)
-        {
-            return;
-        }
-
-        var refreshTask = RefreshRecentFolderAvailabilityAsync(lifetime.Token);
-        _recentFolderAvailabilityRefreshTask = refreshTask;
-        ObserveDetachedTask(refreshTask, "RefreshRecentFolderAvailability");
-    }
-
-    private async Task RefreshRecentFolderAvailabilityAsync(CancellationToken cancellationToken)
-    {
-        var paths = _viewModel.RecentFolders
-            .Select(static folder => folder.Value)
-            .ToArray();
-        if (paths.Length == 0)
-        {
-            _unavailableRecentFolderPaths.Clear();
-            return;
-        }
-
-        var availability = await _recentFolderAvailabilityService.CheckAsync(paths, cancellationToken);
-        cancellationToken.ThrowIfCancellationRequested();
-
-        _unavailableRecentFolderPaths.IntersectWith(paths);
-        foreach (var (path, isAvailable) in availability)
-            UpdateRecentFolderAvailability(path, isAvailable);
-
-        ApplyRecentFolderAvailabilityToMenu();
-    }
-
-    private void ApplyRecentFolderAvailabilityToMenu()
-    {
-        if (_topMenuBar?.RecentMenuItemControl is not { } recentMenuItem)
-            return;
-
-        foreach (var item in recentMenuItem.Items.OfType<MenuItem>())
-        {
-            if (item.Tag is string path)
-            {
-                SetRecentFolderMenuItemAvailability(
-                    item,
-                    !_unavailableRecentFolderPaths.Contains(path));
-            }
-        }
-    }
-
-    private void UpdateRecentFolderAvailability(string path, bool isAvailable)
-    {
-        if (isAvailable)
-            _unavailableRecentFolderPaths.Remove(path);
-        else
-            _unavailableRecentFolderPaths.Add(path);
-    }
-
-    private static void SetRecentFolderMenuItemAvailability(MenuItem item, bool isAvailable)
-    {
-        const string unavailableClass = "recent-folder-unavailable";
-        if (isAvailable)
-            item.Classes.Remove(unavailableClass);
-        else if (!item.Classes.Contains(unavailableClass))
-            item.Classes.Add(unavailableClass);
-    }
-
-    private void RemoveRecentFolder(string path)
-    {
-        _recentProjectsDb = _recentProjectsStore.RemoveFolder(_recentProjectsDb, path);
-        _unavailableRecentFolderPaths.Remove(path);
-        SyncRecentProjectsToViewModel();
-        RefreshRecentFoldersMenu();
-    }
-
-    private void RecordRecentFolder(string path)
-    {
-        _recentProjectsDb = _recentProjectsStore.AddFolder(_recentProjectsDb, path);
-        _unavailableRecentFolderPaths.Remove(path);
-        SyncRecentProjectsToViewModel();
-        RefreshRecentFoldersMenu();
-    }
-
-    private void RecordRecentRepository(string repositoryUrl)
-    {
-        _recentProjectsDb = _recentProjectsStore.AddRepository(_recentProjectsDb, repositoryUrl);
-        SyncRecentProjectsToViewModel();
-    }
-
-    #endregion
-
-    #region Language Menu
-
-    private void RefreshLanguageMenuChecks()
-    {
-        foreach (var (item, language, label) in EnumerateLanguageMenuItems())
-        {
-            if (item is null)
-                continue;
-
-            item.Header = CreateCheckedMenuHeader(_localization.CurrentLanguage == language, label);
-        }
-    }
-
-    private IEnumerable<(MenuItem? Item, AppLanguage Language, string Label)> EnumerateLanguageMenuItems()
-    {
-        var topMenuBar = _topMenuBar;
-        if (topMenuBar is null)
-            yield break;
-
-        yield return (topMenuBar.LanguageEnMenuItemControl, AppLanguage.En, "English");
-        yield return (topMenuBar.LanguageRuMenuItemControl, AppLanguage.Ru, "Русский");
-        yield return (topMenuBar.LanguageEsMenuItemControl, AppLanguage.Es, "Español");
-        yield return (topMenuBar.LanguagePtMenuItemControl, AppLanguage.Pt, "Português (Brasil)");
-        yield return (topMenuBar.LanguagePtPtMenuItemControl, AppLanguage.PtPt, "Português (Portugal)");
-        yield return (topMenuBar.LanguageDeMenuItemControl, AppLanguage.De, "Deutsch");
-        yield return (topMenuBar.LanguageFrMenuItemControl, AppLanguage.Fr, "Français");
-        yield return (topMenuBar.LanguageItMenuItemControl, AppLanguage.It, "Italiano");
-        yield return (topMenuBar.LanguageTgMenuItemControl, AppLanguage.Tg, "Тоҷикӣ");
-        yield return (topMenuBar.LanguageUzMenuItemControl, AppLanguage.Uz, "Oʻzbek");
-        yield return (topMenuBar.LanguageKkMenuItemControl, AppLanguage.Kk, "Қазақ");
-    }
-
-    private static string CreateCheckedMenuHeader(bool isChecked, string label)
-        => isChecked ? $"✓ {label}" : $"   {label}";
-
-    #endregion
-
-    #region Tree Font Menu
-
-    private void AttachTreeFontMenuHandlers()
-    {
-        if (_topMenuBar?.TreeFontMenuItemControl is { } treeFontMenuItem)
-            treeFontMenuItem.SubmenuOpened += OnTreeFontMenuSubmenuOpened;
-    }
-
-    private void DetachTreeFontMenuHandlers()
-    {
-        if (_topMenuBar?.TreeFontMenuItemControl is { } treeFontMenuItem)
-            treeFontMenuItem.SubmenuOpened -= OnTreeFontMenuSubmenuOpened;
-    }
-
-    private void OnTreeFontMenuSubmenuOpened(object? sender, RoutedEventArgs e)
-    {
-        RefreshTreeFontMenu();
-    }
-
-    private void RefreshTreeFontMenu()
-    {
-        var treeFontMenuItem = _topMenuBar?.TreeFontMenuItemControl;
-        if (treeFontMenuItem is null)
-            return;
-
-        treeFontMenuItem.Items.Clear();
-        foreach (var fontFamily in EnumerateTreeFontMenuFamilies())
-            treeFontMenuItem.Items.Add(CreateTreeFontMenuItem(fontFamily));
-    }
-
-    private IEnumerable<FontFamily> EnumerateTreeFontMenuFamilies()
-    {
-        var seen = new HashSet<string>(StringComparer.OrdinalIgnoreCase);
-        foreach (var fontFamily in _viewModel.FontFamilies)
-        {
-            if (seen.Add(GetTreeFontKey(fontFamily)))
-                yield return fontFamily;
-        }
-    }
-
-    private MenuItem CreateTreeFontMenuItem(FontFamily fontFamily)
-    {
-        var displayName = GetTreeFontDisplayName(fontFamily);
-        var item = new MenuItem
-        {
-            Header = CreateCheckedMenuHeader(IsPendingTreeFont(fontFamily), displayName),
-            Tag = fontFamily,
-            MinHeight = TreeFontMenuItemHeight
-        };
-
-        item.Click += OnTreeFontMenuItemClick;
-        return item;
-    }
-
-    private void OnTreeFontMenuItemClick(object? sender, RoutedEventArgs e)
-    {
-        if (sender is not MenuItem { Tag: FontFamily fontFamily })
-            return;
-
-        _viewModel.PendingFontFamily = fontFamily;
-        e.Handled = true;
-    }
-
-    private bool IsPendingTreeFont(FontFamily fontFamily)
-        => AreSameTreeFont(_viewModel.PendingFontFamily, fontFamily);
-
-    private string GetTreeFontDisplayName(FontFamily fontFamily)
-    {
-        if (IsDefaultTreeFont(fontFamily))
-            return _viewModel.SettingsFontDefault;
-
-        var name = fontFamily.Name?.Trim();
-        return string.IsNullOrWhiteSpace(name) ? _viewModel.SettingsFontDefault : name;
-    }
-
-    private static bool AreSameTreeFont(FontFamily? left, FontFamily? right)
-    {
-        if (IsDefaultTreeFont(left) && IsDefaultTreeFont(right))
-            return true;
-
-        return string.Equals(left?.Name, right?.Name, StringComparison.OrdinalIgnoreCase);
-    }
-
-    private static string GetTreeFontKey(FontFamily fontFamily)
-        => IsDefaultTreeFont(fontFamily) ? string.Empty : fontFamily.Name ?? string.Empty;
-
-    private static bool IsDefaultTreeFont(FontFamily? fontFamily)
-    {
-        var name = fontFamily?.Name;
-        return string.IsNullOrWhiteSpace(name) || name.StartsWith("$", StringComparison.Ordinal);
-    }
-
-    #endregion
-
-    #region Git Operations
-
-    private void OnGitClone(object? sender, RoutedEventArgs e)
-    {
-        if (!_viewModel.CanChangeProjectTree)
-            return;
-
-        _viewModel.GitCloneUrl = string.Empty;
-        _viewModel.GitCloneStatus = string.Empty;
-        _viewModel.GitCloneInProgress = false;
-
-        // Create and show Git Clone window
-        _gitCloneWindow = new GitCloneWindow
-        {
-            DataContext = _viewModel
-        };
-
-        _gitCloneWindow.StartCloneRequested += OnGitCloneStart;
-        _gitCloneWindow.CancelRequested += OnGitCloneCancel;
-
-        _gitCloneWindow.ShowDialog(this);
-        e.Handled = true;
-    }
-
-    private void OnGitCloneClose(object? sender, RoutedEventArgs e)
-    {
-        CancelGitCloneOperation();
-        _gitCloneWindow?.Close();
-        _gitCloneWindow = null;
-        e.Handled = true;
-    }
-
-    private async void OnGitCloneStart(object? sender, RoutedEventArgs e)
-    {
-        var url = _viewModel.GitCloneUrl?.Trim();
-        if (string.IsNullOrWhiteSpace(url))
-        {
-            await ShowErrorAsync(_viewModel.GitErrorInvalidUrl);
-            return;
-        }
-
-        // Validate URL format before attempting to clone
-        if (!IsValidGitRepositoryUrl(url))
-        {
-            await ShowErrorAsync(_viewModel.GitErrorInvalidUrl);
-            return;
-        }
-
-        var gitCloneCts = ReplaceCancellationSource(ref _gitCloneCts);
-        var cancellationToken = gitCloneCts.Token;
-
-        _viewModel.GitCloneInProgress = true;
-        _viewModel.GitCloneStatus = _viewModel.GitCloneProgressCheckingGit;
-        _taskbarProgress.BeginGitClone();
-
-        string? targetPath = null;
-
-        try
-        {
-            // Check internet connection before starting
-            var hasInternet = await CheckInternetConnectionAsync(cancellationToken);
-            if (!hasInternet)
-            {
-                _viewModel.GitCloneInProgress = false;
-                _gitCloneWindow?.Close();
-                _gitCloneWindow = null;
-                _taskbarProgress.MarkGitCloneError();
-                await ShowErrorAsync(_viewModel.GitErrorNoInternetConnection);
-                return;
-            }
-
-            // Clean up previous cached repository before cloning a new one
-            if (_currentCachedRepoPath is not null)
-            {
-                _repoCacheService.DeleteRepositoryDirectory(_currentCachedRepoPath);
-                _currentCachedRepoPath = null;
-            }
-
-            targetPath = _repoCacheService.CreateRepositoryDirectory(url);
-
-            // Track current operation for progress reporting
-            string currentOperation = string.Empty;
-
-            var progress = new Progress<string>(status =>
-            {
-                Dispatcher.Post(() =>
-                {
-                    _taskbarProgress.UpdateGitClone(status);
-
-                    // Handle phase transition markers
-                    if (status == "::EXTRACTING::")
-                    {
-                        currentOperation = _viewModel.GitCloneProgressExtracting;
-                        _viewModel.GitCloneStatus = currentOperation;
-                        return;
-                    }
-
-                    // Keep localized phase labels and append numeric progress only.
-                    // Raw git stderr lines (e.g. "Cloning into ...") are not shown in UI.
-                    if (status.EndsWith('%') && status.Length <= 4 && !string.IsNullOrEmpty(currentOperation))
-                    {
-                        _viewModel.GitCloneStatus = $"{currentOperation} {status}";
-                    }
-                    else if (!string.IsNullOrEmpty(currentOperation))
-                    {
-                        _viewModel.GitCloneStatus = currentOperation;
-                    }
-                });
-            });
-
-            GitCloneResult result;
-
-            // Check if Git is available
-            var gitAvailable = await _gitService.IsGitAvailableAsync(cancellationToken);
-
-            if (gitAvailable)
-            {
-                currentOperation = _viewModel.GitCloneProgressCloning;
-                _viewModel.GitCloneStatus = currentOperation;
-                _taskbarProgress.SetGitCloneIndeterminate();
-                result = await _gitService.CloneAsync(url, targetPath, progress, cancellationToken);
-            }
-            else
-            {
-                // Fallback to ZIP download
-                _viewModel.GitCloneStatus = _viewModel.GitErrorGitNotFound;
-                await Task.Delay(1500, cancellationToken);
-
-                currentOperation = _viewModel.GitCloneProgressDownloading;
-                _viewModel.GitCloneStatus = currentOperation;
-                _taskbarProgress.SetGitCloneIndeterminate();
-                result = await _zipDownloadService.DownloadAndExtractAsync(url, targetPath, progress, cancellationToken);
-            }
-
-            cancellationToken.ThrowIfCancellationRequested();
-
-            if (!result.Success)
-            {
-                _repoCacheService.DeleteRepositoryDirectory(targetPath);
-                _gitCloneWindow?.Close();
-                _gitCloneWindow = null;
-                _viewModel.GitCloneInProgress = false;
-                _taskbarProgress.MarkGitCloneError();
-                await ShowErrorAsync(_localization.Format("Git.Error.CloneFailed", result.ErrorMessage ?? "Unknown error"));
-                _toastService.Show(_localization["Toast.Git.CloneError"]);
-                return;
-            }
-
-            await ApplySuccessfulGitCloneAsync(result, targetPath, url, cancellationToken);
-        }
-        catch (OperationCanceledException)
-        {
-            if (targetPath is not null)
-            {
-                // Use default cancellation token since operation was cancelled
-                _repoCacheService.DeleteRepositoryDirectory(targetPath);
-            }
-        }
-        catch (Exception ex)
-        {
-            if (targetPath is not null)
-            {
-                _repoCacheService.DeleteRepositoryDirectory(targetPath);
-            }
-
-            _gitCloneWindow?.Close();
-            _gitCloneWindow = null;
-            _taskbarProgress.MarkGitCloneError();
-            await ShowErrorAsync(_localization.Format("Git.Error.CloneFailed", ex.Message));
-            _toastService.Show(_localization["Toast.Git.CloneError"]);
-        }
-        finally
-        {
-            _viewModel.GitCloneInProgress = false;
-            _taskbarProgress.CompleteGitClone();
-            DisposeIfCurrent(ref _gitCloneCts, gitCloneCts);
-        }
-
-        e.Handled = true;
-    }
-
-    internal async Task ApplySuccessfulGitCloneAsync(
-        GitCloneResult result,
-        string cachePath,
-        string requestedUrl,
-        CancellationToken cancellationToken)
-    {
-        ArgumentNullException.ThrowIfNull(result);
-        cancellationToken.ThrowIfCancellationRequested();
-
-        _gitCloneWindow?.Close();
-        _gitCloneWindow = null;
-        _viewModel.GitCloneInProgress = false;
-        _viewModel.ProjectSourceType = result.SourceType;
-        _viewModel.CurrentBranch = result.DefaultBranch ?? "main";
-        _currentProjectDisplayName = result.RepositoryName;
-        _currentRepositoryUrl = result.RepositoryUrl;
-        _currentCachedRepoPath = cachePath;
-
-        var opened = await TryOpenFolderAsync(result.LocalPath, fromDialog: false, recordRecentFolder: false);
-        if (!opened || !PathComparer.Default.Equals(_currentPath, result.LocalPath))
-            return;
-
-        RecordRecentRepository(string.IsNullOrWhiteSpace(result.RepositoryUrl) ? requestedUrl : result.RepositoryUrl);
-
-        // Clone-only branch discovery stays behind the reveal barrier so it cannot compete
-        // with the settings animation or make the clone completion appear frozen.
-        if (result.SourceType == ProjectSourceType.GitClone)
-        {
-            var visualReadyTask = _postLoadVisualReadyTask;
-            await MetricsCalculationPolicy.WaitForInitialVisualReadyAsync(
-                visualReadyTask,
-                MetricsCalculationPolicy.InitialVisualReadyTimeout,
-                cancellationToken);
-
-            if (PathComparer.Default.Equals(_currentPath, result.LocalPath))
-                await RefreshGitBranchesAsync(result.LocalPath, cancellationToken);
-        }
-
-        if (PathComparer.Default.Equals(_currentPath, result.LocalPath))
-            _toastService.Show(_localization["Toast.Git.CloneSuccess"]);
-    }
-
-    private void OnGitCloneCancel(object? sender, RoutedEventArgs e)
-    {
-        if (_viewModel.GitCloneInProgress)
-        {
-            CancelGitCloneOperation();
-        }
-        else
-        {
-            _gitCloneWindow?.Close();
-            _gitCloneWindow = null;
-        }
-        e.Handled = true;
-    }
-
-    private void CancelGitCloneOperation()
-    {
-        _gitCloneCts?.Cancel();
-        _viewModel.GitCloneInProgress = false;
-        _taskbarProgress.CompleteGitClone();
-    }
-
-    private async void OnGitGetUpdates(object? sender, RoutedEventArgs e)
-    {
-        if (!_viewModel.CanGetGitUpdates)
-            return;
-
-        await GetGitUpdatesAsync();
-        e.Handled = true;
-    }
-
-    private async Task GetGitUpdatesAsync()
-    {
-        if (!_viewModel.IsGitMode || string.IsNullOrEmpty(_currentPath))
-            return;
-
-        var gitCts = ReplaceCancellationSource(ref _gitOperationCts);
-        var cancellationToken = gitCts.Token;
-        long? statusOperationId = null;
-        try
-        {
-            var statusText = string.IsNullOrWhiteSpace(_viewModel.CurrentBranch)
-                ? _viewModel.StatusOperationGettingUpdates
-                : _localization.Format("Status.Operation.GettingUpdatesBranch", _viewModel.CurrentBranch);
-            statusOperationId = _statusOperations.Begin(
-                statusText,
-                indeterminate: true,
-                operationType: StatusOperationType.GitPullUpdates,
-                cancelAction: () => gitCts.Cancel());
-
-            var progress = new Progress<string>(status =>
-            {
-                Dispatcher.Post(() =>
-                {
-                    if (GitProgressStatusParser.TryParseTrailingPercent(status, out var percent))
-                        _statusOperations.UpdateProgress(percent, statusText, statusOperationId);
-                    else
-                        _statusOperations.UpdateText(statusText, statusOperationId);
-                });
-            });
-            var beforeHash = await _gitService.GetHeadCommitAsync(_currentPath, cancellationToken);
-            var success = await _gitService.PullUpdatesAsync(_currentPath, progress, cancellationToken);
-
-            if (!success)
-            {
-                _statusOperations.Complete(statusOperationId);
-                await ShowErrorAsync(_localization.Format("Git.Error.UpdateFailed", "Pull failed"));
-                return;
-            }
-
-            // Refresh branches and tree
-            await RefreshGitBranchesAsync(_currentPath, cancellationToken);
-            await ReloadProjectAsync(cancellationToken);
-
-            var afterHash = await _gitService.GetHeadCommitAsync(_currentPath, cancellationToken);
-            if (!string.IsNullOrWhiteSpace(beforeHash) && !string.IsNullOrWhiteSpace(afterHash) && beforeHash == afterHash)
-            {
-                _toastService.Show(_localization["Toast.Git.NoUpdates"]);
-                _statusOperations.Complete(statusOperationId);
-            }
-            else
-            {
-                _toastService.Show(_localization["Toast.Git.UpdatesApplied"]);
-                _statusOperations.Complete(statusOperationId);
-                // Clean up memory from old tree after successful update.
-                ScheduleBackgroundMemoryCleanup(MemoryCleanupReason.GitPullUpdate);
-            }
-        }
-        catch (OperationCanceledException) when (cancellationToken.IsCancellationRequested)
-        {
-            _statusOperations.Complete(statusOperationId);
-            _toastService.Show(_localization["Toast.Operation.GitCanceled"]);
-        }
-        catch (Exception ex)
-        {
-            _statusOperations.Complete(statusOperationId);
-            await ShowErrorAsync(_localization.Format("Git.Error.UpdateFailed", ex.Message));
-        }
-        finally
-        {
-            DisposeIfCurrent(ref _gitOperationCts, gitCts);
-        }
-    }
-
-    private async void OnGitBranchSwitch(object? sender, string branchName)
-    {
-        if (!_viewModel.CanGetGitUpdates || string.IsNullOrEmpty(_currentPath))
-            return;
-
-        var gitCts = ReplaceCancellationSource(ref _gitOperationCts);
-        var cancellationToken = gitCts.Token;
-        long? statusOperationId = null;
-        try
-        {
-            var statusText = _localization.Format("Status.Operation.SwitchingBranch", branchName);
-            statusOperationId = _statusOperations.Begin(
-                statusText,
-                indeterminate: true,
-                operationType: StatusOperationType.GitSwitchBranch,
-                cancelAction: () => gitCts.Cancel());
-
-            var progress = new Progress<string>(status =>
-            {
-                Dispatcher.Post(() =>
-                {
-                    if (GitProgressStatusParser.TryParseTrailingPercent(status, out var percent))
-                        _statusOperations.UpdateProgress(percent, statusText, statusOperationId);
-                    else
-                        _statusOperations.UpdateText(statusText, statusOperationId);
-                });
-            });
-            var success = await _gitService.SwitchBranchAsync(_currentPath, branchName, progress, cancellationToken);
-
-            // A lightweight retry helps recover from transient remote/network hiccups.
-            if (!success)
-                success = await _gitService.SwitchBranchAsync(_currentPath, branchName, progress: null, cancellationToken);
-
-            if (!success)
-            {
-                _statusOperations.Complete(statusOperationId);
-                await ShowErrorAsync(_localization.Format("Git.Error.BranchSwitchFailed", branchName));
-                return;
-            }
-
-            // Reload tree first so branch/title state is only updated after full success.
-            // This keeps UI stable if reload fails or gets cancelled mid-flight.
-            await ReloadProjectAsync(cancellationToken);
-            await RefreshGitBranchesAsync(_currentPath, cancellationToken);
-            _statusOperations.Complete(statusOperationId);
-
-            _viewModel.CurrentBranch = branchName;
-            UpdateTitle();
-            _toastService.Show(_localization.Format("Toast.Git.BranchSwitched", branchName));
-
-            // Clean up memory from old branch tree.
-            ScheduleBackgroundMemoryCleanup(MemoryCleanupReason.GitBranchSwitch);
-        }
-        catch (OperationCanceledException) when (cancellationToken.IsCancellationRequested)
-        {
-            _statusOperations.Complete(statusOperationId);
-            _toastService.Show(_localization["Toast.Operation.GitCanceled"]);
-        }
-        catch (Exception ex)
-        {
-            _statusOperations.Complete(statusOperationId);
-            await ShowErrorAsync(_localization.Format("Git.Error.BranchSwitchFailed", ex.Message));
-        }
-        finally
-        {
-            DisposeIfCurrent(ref _gitOperationCts, gitCts);
-        }
-    }
-
-    private async Task RefreshGitBranchesAsync(string repositoryPath, CancellationToken cancellationToken = default)
-    {
-        try
-        {
-            var branches = await _gitService.GetBranchesAsync(repositoryPath, cancellationToken);
-
-            _viewModel.GitBranches.Clear();
-            foreach (var branch in branches)
-                _viewModel.GitBranches.Add(branch);
-
-            // Update branch menu
-            UpdateBranchMenu();
-        }
-        catch (OperationCanceledException)
-        {
-            throw;
-        }
-        catch
-        {
-            // Ignore branch loading errors
-        }
-    }
-
-    private void UpdateBranchMenu()
-    {
-        var branchMenuItem = _topMenuBar?.GitBranchMenuItemControl;
-        if (branchMenuItem is null)
-            return;
-
-        // Clear old items - they will be garbage collected since they have no external references
-        // and we're using a named handler method instead of lambda captures
-        branchMenuItem.Items.Clear();
-        GitBranchMenuScrollBehavior.SetScrollable(branchMenuItem, _viewModel.GitBranches.Count);
-
-        foreach (var branch in _viewModel.GitBranches)
-            branchMenuItem.Items.Add(CreateBranchMenuItem(branch));
-    }
-
-    private MenuItem CreateBranchMenuItem(GitBranch branch)
-    {
-        var item = new MenuItem
-        {
-            Header = CreateCheckedMenuHeader(branch.IsActive, branch.Name),
-            Tag = branch.Name,
-            MinHeight = BranchMenuItemHeight
-        };
-
-        // Use a named handler to avoid closure captures and keep menu rebuilds cheap.
-        item.Click += OnBranchMenuItemClick;
-        return item;
-    }
-
-    private void OnBranchMenuItemClick(object? sender, RoutedEventArgs e)
-    {
-        if (_viewModel.CanChangeProjectTree && sender is MenuItem { Tag: string name })
-            _topMenuBar?.OnGitBranchSwitch(name);
-    }
-
-    #endregion
 
     private void OnAboutOpenLink(object? sender, RoutedEventArgs e)
     {
@@ -8206,250 +6939,9 @@ public partial class MainWindow : Window
         }
     }
 
-    private bool TryBuildInteractiveFilteredTreeResult(
-        string? nameFilter,
-        CancellationToken cancellationToken,
-        out BuildTreeResult result)
-    {
-        result = default!;
-        var baseTree = _filterBaseTree;
-        if (baseTree is null)
-            return false;
-
-        cancellationToken.ThrowIfCancellationRequested();
-        if (string.IsNullOrWhiteSpace(nameFilter))
-        {
-            ResetInteractiveFilterCache();
-            result = baseTree;
-            return true;
-        }
-
-        if (TryGetCachedInteractiveFilterRoot(nameFilter, out var cachedRoot))
-        {
-            _lastInteractiveFilterBaseRoot = baseTree.Root;
-            _lastInteractiveFilteredRoot = cachedRoot;
-            _lastInteractiveFilterQuery = nameFilter;
-            result = new BuildTreeResult(
-                Root: cachedRoot,
-                RootAccessDenied: baseTree.RootAccessDenied,
-                HadAccessDenied: baseTree.HadAccessDenied);
-            return true;
-        }
-
-        // For incremental typing, prefer the narrowest known prefix source.
-        // This reduces traversal work while preserving correctness.
-        var filterSourceRoot = baseTree.Root;
-        if (_lastInteractiveFilterBaseRoot is not null &&
-            ReferenceEquals(_lastInteractiveFilterBaseRoot, baseTree.Root))
-        {
-            if (_lastInteractiveFilteredRoot is not null &&
-                !string.IsNullOrWhiteSpace(_lastInteractiveFilterQuery) &&
-                nameFilter.StartsWith(_lastInteractiveFilterQuery, StringComparison.OrdinalIgnoreCase))
-            {
-                filterSourceRoot = _lastInteractiveFilteredRoot;
-            }
-            else if (TryGetBestInteractiveFilterPrefixSource(nameFilter, out var prefixRoot))
-            {
-                filterSourceRoot = prefixRoot;
-            }
-        }
-
-        var filteredRoot = FilterTreeForNameQuery(filterSourceRoot, nameFilter, cancellationToken);
-        _lastInteractiveFilterBaseRoot = baseTree.Root;
-        _lastInteractiveFilteredRoot = filteredRoot;
-        _lastInteractiveFilterQuery = nameFilter;
-        CacheInteractiveFilterRoot(nameFilter, filteredRoot);
-
-        result = new BuildTreeResult(
-            Root: filteredRoot,
-            RootAccessDenied: baseTree.RootAccessDenied,
-            HadAccessDenied: baseTree.HadAccessDenied);
-        return true;
-    }
-
-    private static TreeNodeDescriptor FilterTreeForNameQuery(
-        TreeNodeDescriptor root,
-        string query,
-        CancellationToken cancellationToken)
-    {
-        List<TreeNodeDescriptor>? filteredChildren = null;
-        var originalChildren = root.Children;
-
-        for (var index = 0; index < originalChildren.Count; index++)
-        {
-            cancellationToken.ThrowIfCancellationRequested();
-            var originalChild = originalChildren[index];
-            var filteredChild = FilterTreeNodeByName(originalChild, query, cancellationToken);
-
-            if (filteredChild is null)
-            {
-                if (filteredChildren is null)
-                {
-                    filteredChildren = new List<TreeNodeDescriptor>(Math.Min(originalChildren.Count, 16));
-                    for (var j = 0; j < index; j++)
-                        filteredChildren.Add(originalChildren[j]);
-                }
-
-                continue;
-            }
-
-            if (filteredChildren is not null)
-            {
-                filteredChildren.Add(filteredChild);
-                continue;
-            }
-
-            if (!ReferenceEquals(filteredChild, originalChild))
-            {
-                filteredChildren = new List<TreeNodeDescriptor>(Math.Min(originalChildren.Count, 16));
-                for (var j = 0; j < index; j++)
-                    filteredChildren.Add(originalChildren[j]);
-                filteredChildren.Add(filteredChild);
-            }
-        }
-
-        if (filteredChildren is null)
-            return root;
-
-        return root with { Children = filteredChildren };
-    }
-
     [MethodImpl(MethodImplOptions.AggressiveInlining)]
-    private void ResetInteractiveFilterCache()
-    {
-        _lastInteractiveFilteredRoot = null;
-        _lastInteractiveFilterBaseRoot = null;
-        _lastInteractiveFilterQuery = null;
-        _interactiveFilterQueryCache.Clear();
-        _interactiveFilterQueryCacheLru.Clear();
-        _interactiveFilterQueryCacheNodes.Clear();
-    }
-
-    private bool TryGetCachedInteractiveFilterRoot(string query, out TreeNodeDescriptor root)
-    {
-        if (_interactiveFilterQueryCache.TryGetValue(query, out root!))
-        {
-            if (_interactiveFilterQueryCacheNodes.TryGetValue(query, out var node))
-            {
-                _interactiveFilterQueryCacheLru.Remove(node);
-                _interactiveFilterQueryCacheLru.AddFirst(node);
-            }
-
-            return true;
-        }
-
-        root = null!;
-        return false;
-    }
-
-    private bool TryGetBestInteractiveFilterPrefixSource(string query, out TreeNodeDescriptor root)
-    {
-        root = null!;
-        string? bestPrefix = null;
-
-        foreach (var cachedQuery in _interactiveFilterQueryCache.Keys)
-        {
-            if (string.IsNullOrWhiteSpace(cachedQuery))
-                continue;
-
-            if (!query.StartsWith(cachedQuery, StringComparison.OrdinalIgnoreCase))
-                continue;
-
-            if (bestPrefix is null || cachedQuery.Length > bestPrefix.Length)
-                bestPrefix = cachedQuery;
-        }
-
-        if (bestPrefix is null)
-            return false;
-
-        return TryGetCachedInteractiveFilterRoot(bestPrefix, out root);
-    }
-
-    private void CacheInteractiveFilterRoot(string query, TreeNodeDescriptor root)
-    {
-        if (string.IsNullOrWhiteSpace(query))
-            return;
-
-        _interactiveFilterQueryCache[query] = root;
-
-        if (_interactiveFilterQueryCacheNodes.TryGetValue(query, out var existingNode))
-        {
-            _interactiveFilterQueryCacheLru.Remove(existingNode);
-            _interactiveFilterQueryCacheLru.AddFirst(existingNode);
-            return;
-        }
-
-        var node = new LinkedListNode<string>(query);
-        _interactiveFilterQueryCacheLru.AddFirst(node);
-        _interactiveFilterQueryCacheNodes[query] = node;
-
-        while (_interactiveFilterQueryCacheNodes.Count > InteractiveFilterQueryCacheLimit)
-        {
-            var last = _interactiveFilterQueryCacheLru.Last;
-            if (last is null)
-                break;
-
-            _interactiveFilterQueryCacheLru.RemoveLast();
-            _interactiveFilterQueryCacheNodes.Remove(last.Value);
-            _interactiveFilterQueryCache.Remove(last.Value);
-        }
-    }
-
-    private static TreeNodeDescriptor? FilterTreeNodeByName(
-        TreeNodeDescriptor node,
-        string query,
-        CancellationToken cancellationToken)
-    {
-        cancellationToken.ThrowIfCancellationRequested();
-
-        var selfMatches = node.DisplayName.Contains(query, StringComparison.OrdinalIgnoreCase);
-        if (node.Children.Count == 0)
-            return selfMatches ? node : null;
-
-        List<TreeNodeDescriptor>? filteredChildren = null;
-        var originalChildren = node.Children;
-        var matchedChildrenCount = 0;
-
-        for (var index = 0; index < originalChildren.Count; index++)
-        {
-            var originalChild = originalChildren[index];
-            var filteredChild = FilterTreeNodeByName(originalChild, query, cancellationToken);
-            if (filteredChild is null)
-            {
-                if (filteredChildren is null)
-                {
-                    filteredChildren = new List<TreeNodeDescriptor>(Math.Min(originalChildren.Count, 8));
-                    for (var j = 0; j < index; j++)
-                        filteredChildren.Add(originalChildren[j]);
-                }
-
-                continue;
-            }
-
-            matchedChildrenCount++;
-            if (filteredChildren is not null)
-            {
-                filteredChildren.Add(filteredChild);
-                continue;
-            }
-
-            if (!ReferenceEquals(filteredChild, originalChild))
-            {
-                filteredChildren = new List<TreeNodeDescriptor>(Math.Min(originalChildren.Count, 8));
-                for (var j = 0; j < index; j++)
-                    filteredChildren.Add(originalChildren[j]);
-                filteredChildren.Add(filteredChild);
-            }
-        }
-
-        if (!selfMatches && matchedChildrenCount == 0)
-            return null;
-
-        if (filteredChildren is null)
-            return node;
-
-        return node with { Children = filteredChildren };
-    }
+    private void ResetInteractiveFilterCache() =>
+        _refreshPipeline.InvalidateInteractiveFilterCache();
 
     private Task<TreeRefreshOutcome> RefreshTreeAsync(
         bool interactiveFilter = false,
@@ -8976,21 +7468,6 @@ public partial class MainWindow : Window
     private HashSet<string> GetCheckedPaths()
         => _treeSelectionSnapshotCache.GetOrCreate(_viewModel.TreeNodes);
 
-    private IReadOnlyList<string> BuildOrderedUniqueFilePaths(IReadOnlySet<string> selectedPaths)
-    {
-        if (selectedPaths.Count > 0)
-        {
-            if (_currentTree is null)
-                return [];
-
-            return BuildOrderedSelectedFilePaths(_currentTree.Root, selectedPaths);
-        }
-
-        return _currentTree is null
-            ? []
-            : _currentTree.OrderedFilePaths ?? _metrics.GetOrBuildAllOrderedFilePaths(_currentTree.Root);
-    }
-
     private static List<string> BuildOrderedSelectedFilePaths(
         TreeNodeDescriptor treeRoot,
         IReadOnlySet<string> selectedPaths,
@@ -9129,282 +7606,5 @@ public partial class MainWindow : Window
         }
     }
 
-    #region Real-time Status Metrics
-
-    private bool IsBackgroundMetricsActive() => _metrics.IsBackgroundActive;
-
-    private void OnTreeNodeCheckedChanged(TreeNodeViewModel _)
-    {
-        _treeSelectionSnapshotCache.Invalidate();
-        _metrics.ScheduleRecalculate();
-        SchedulePreviewRefresh();
-    }
-
-    private void RenderPreviewSelectionMetrics()
-    {
-        if (!_hasPreviewSelectionMetricsSnapshot)
-        {
-            _viewModel.StatusPreviewSelectionVisible = false;
-            _viewModel.StatusPreviewSelectionStatsText = string.Empty;
-            return;
-        }
-
-        _viewModel.StatusPreviewSelectionStatsText = PreviewSelectionMetricsPolicy.FormatStatusMetricsText(
-            _lastPreviewSelectionMetrics,
-            BuildStatusMetricLabels(),
-            useCompactMode: false);
-        _viewModel.StatusPreviewSelectionVisible = true;
-    }
-
-    private StatusMetricLabels BuildStatusMetricLabels()
-    {
-        var linesLabel = _localization.Format("Status.Metric.Lines", "{0}");
-        var charsLabel = _localization.Format("Status.Metric.Chars", "{0}");
-        var tokensLabel = _localization.Format("Status.Metric.Tokens", "{0}");
-
-        return new StatusMetricLabels(
-            linesLabel.Replace("{0}", string.Empty).Trim(),
-            charsLabel.Replace("{0}", string.Empty).Trim(),
-            tokensLabel.Replace("{0}", string.Empty).Trim());
-    }
-
-    private void SchedulePreviewSelectionMetricsUpdate(bool immediate = false)
-    {
-        if (!_viewModel.IsAnyPreviewVisible || _previewTextControl is null)
-        {
-            ClearPreviewSelectionMetrics();
-            return;
-        }
-
-        if (!_previewTextControl.TryGetSelectionRange(out _))
-        {
-            ClearPreviewSelectionMetrics();
-            return;
-        }
-
-        if (immediate)
-        {
-            _previewSelectionMetricsDebounceTimer?.Stop();
-            RecalculatePreviewSelectionMetricsAsync();
-            return;
-        }
-
-        if (_previewSelectionMetricsDebounceTimer is null)
-        {
-            _previewSelectionMetricsDebounceTimer = new DispatcherTimer(DispatcherPriority.Background, Dispatcher)
-            {
-                Interval = PreviewSelectionMetricsDebounceInterval
-            };
-            _previewSelectionMetricsDebounceTimer.Tick += OnPreviewSelectionMetricsDebounceTick;
-        }
-
-        _previewSelectionMetricsDebounceTimer.Stop();
-        _previewSelectionMetricsDebounceTimer.Start();
-    }
-
-    private void OnPreviewSelectionMetricsDebounceTick(object? sender, EventArgs e)
-    {
-        _previewSelectionMetricsDebounceTimer?.Stop();
-        RecalculatePreviewSelectionMetricsAsync();
-    }
-
-    private void RecalculatePreviewSelectionMetricsAsync()
-    {
-        if (!TryCapturePreviewSelectionMetricsSnapshot(out var snapshot))
-        {
-            ClearPreviewSelectionMetrics();
-            return;
-        }
-
-        if (TryGetCachedPreviewSelectionMetrics(snapshot, out var cachedMetrics))
-        {
-            _previewSelectionMetricsDebounceTimer?.Stop();
-            var previousCts = Interlocked.Exchange(ref _previewSelectionMetricsCts, null);
-            previousCts?.Cancel();
-            previousCts?.Dispose();
-            Interlocked.Increment(ref _previewSelectionMetricsVersion);
-            _lastPreviewSelectionMetrics = cachedMetrics;
-            _hasPreviewSelectionMetricsSnapshot = true;
-            RenderPreviewSelectionMetrics();
-            return;
-        }
-
-        var metricsCts = ReplaceCancellationSource(ref _previewSelectionMetricsCts);
-        var token = metricsCts.Token;
-        var version = Interlocked.Increment(ref _previewSelectionMetricsVersion);
-
-        _ = RecalculatePreviewSelectionMetricsCoreAsync(snapshot, metricsCts, token, version);
-    }
-
-    private async Task RecalculatePreviewSelectionMetricsCoreAsync(
-        PreviewSelectionMetricsSnapshot snapshot,
-        CancellationTokenSource metricsCts,
-        CancellationToken cancellationToken,
-        int version)
-    {
-        try
-        {
-            var metrics = await Task.Run(
-                () => PreviewSelectionMetricsCalculator.Calculate(
-                    snapshot.Document,
-                    snapshot.SelectionRange,
-                    cancellationToken),
-                cancellationToken);
-
-            if (cancellationToken.IsCancellationRequested)
-                return;
-
-            await Dispatcher.UIThread.InvokeAsync(() =>
-            {
-                if (cancellationToken.IsCancellationRequested ||
-                    version != Volatile.Read(ref _previewSelectionMetricsVersion))
-                {
-                    return;
-                }
-
-                if (!TryCapturePreviewSelectionMetricsSnapshot(out var currentSnapshot) ||
-                    !ReferenceEquals(currentSnapshot.Document, snapshot.Document) ||
-                    currentSnapshot.SelectionRange != snapshot.SelectionRange)
-                {
-                    return;
-                }
-
-                _lastPreviewSelectionMetrics = metrics;
-                _hasPreviewSelectionMetricsSnapshot = metrics != ExportOutputMetrics.Empty;
-                RenderPreviewSelectionMetrics();
-            }, DispatcherPriority.Background);
-        }
-        catch (OperationCanceledException)
-        {
-        }
-        catch (ObjectDisposedException)
-        {
-        }
-        finally
-        {
-            DisposeIfCurrent(ref _previewSelectionMetricsCts, metricsCts);
-        }
-    }
-
-    private bool TryCapturePreviewSelectionMetricsSnapshot(out PreviewSelectionMetricsSnapshot snapshot)
-    {
-        snapshot = default;
-
-        if (!_viewModel.IsAnyPreviewVisible || _previewTextControl is null)
-            return false;
-
-        var document = _previewTextControl.Document ?? _viewModel.PreviewDocument;
-        if (document is null)
-            return false;
-
-        if (!_previewTextControl.TryGetSelectionRange(out var selectionRange))
-            return false;
-
-        snapshot = new PreviewSelectionMetricsSnapshot(document, selectionRange);
-        return true;
-    }
-
-    private bool TryGetCachedPreviewSelectionMetrics(
-        PreviewSelectionMetricsSnapshot snapshot,
-        out ExportOutputMetrics metrics)
-    {
-        return _metrics.TryGetCachedPreviewSelectionMetrics(
-            _viewModel.SelectedPreviewContentMode,
-            snapshot.Document,
-            snapshot.SelectionRange,
-            out metrics);
-    }
-
-    private void ClearPreviewSelectionMetrics()
-    {
-        _previewSelectionMetricsDebounceTimer?.Stop();
-        var previousCts = Interlocked.Exchange(ref _previewSelectionMetricsCts, null);
-        previousCts?.Cancel();
-        previousCts?.Dispose();
-        Interlocked.Increment(ref _previewSelectionMetricsVersion);
-
-        _lastPreviewSelectionMetrics = ExportOutputMetrics.Empty;
-        _hasPreviewSelectionMetricsSnapshot = false;
-        _viewModel.StatusPreviewSelectionVisible = false;
-        _viewModel.StatusPreviewSelectionStatsText = string.Empty;
-    }
-
-    private void OnStatusOperationCancelRequested(object? sender, RoutedEventArgs e)
-    {
-        var activeOperation = _statusOperations.GetActiveSnapshot();
-        var activeOperationId = activeOperation.OperationId;
-        var activeOperationType = activeOperation.OperationType;
-
-        // Primary cancellation path for the currently visible status operation.
-        try
-        {
-            activeOperation.CancelAction?.Invoke();
-        }
-        catch
-        {
-            // Ignore cancellation callback errors and continue with fallback logic.
-        }
-
-        // Scoped fallback path: cancel only the currently active operation family.
-        switch (activeOperationType)
-        {
-            case StatusOperationType.LoadProject:
-                _projectLoadPipeline.CancelActiveLoad();
-                break;
-            case StatusOperationType.RefreshProject:
-                _projectOperationCts?.Cancel();
-                _refreshPipeline.CancelActiveRefresh();
-                break;
-            case StatusOperationType.GitPullUpdates:
-            case StatusOperationType.GitSwitchBranch:
-                _gitOperationCts?.Cancel();
-                break;
-            case StatusOperationType.PreviewBuild:
-                _previewPipeline.CancelActiveBuild();
-                break;
-            case StatusOperationType.SelectionRefresh:
-                if (_selectionCoordinator.CancelPendingRefreshes())
-                    _toastService.Show(_localization["Toast.Operation.RefreshCanceled"]);
-                break;
-            case StatusOperationType.ApplySettings:
-                _applySettingsCts?.Cancel();
-                _selectionCoordinator.CancelPendingRefreshes();
-                _refreshPipeline.CancelActiveRefresh();
-                break;
-            case StatusOperationType.ProjectCopyExport:
-                _projectCopyExportCts?.Cancel();
-                break;
-            case StatusOperationType.MetricsCalculation:
-                // Metrics cancellation is handled below by dedicated fallback logic.
-                break;
-            case StatusOperationType.None:
-            default:
-                break;
-        }
-
-        if (activeOperationType == StatusOperationType.MetricsCalculation)
-        {
-            _metrics.CancelByUser();
-            _toastService.Show(_localization["Toast.Operation.MetricsCanceled"]);
-        }
-
-        if (activeOperationType == StatusOperationType.LoadProject)
-        {
-            if (TryApplyActiveProjectLoadCancellationFallback())
-                _toastService.Show(_localization["Toast.Operation.LoadCanceled"]);
-        }
-
-        // Cancel preview build if in progress
-        if (_viewModel.IsPreviewLoading || activeOperationType == StatusOperationType.PreviewBuild)
-        {
-            _previewPipeline.CancelActiveBuild();
-            _viewModel.IsPreviewLoading = false;
-            _toastService.Show(_viewModel.ToastPreviewCanceled);
-        }
-
-        _statusOperations.Complete(activeOperationId);
-    }
-
-    #endregion
 
 }
