+++ Tests/DevProjex.Tests.Unit/ProjectCopyExportUiContractTests.cs
@@ -108,7 +108,7 @@ public sealed class ProjectCopyExportUiContractTests
             "Apps",
             "Avalonia",
             "DevProjex.Avalonia",
-            "MainWindow.axaml.cs");
+            "MainWindow.TextOutput.cs");
         var methodStart = source.IndexOf(
             "private async Task<bool> TryExportTextToFileAsync",
             StringComparison.Ordinal);
@@ -184,6 +184,11 @@ public sealed class ProjectCopyExportUiContractTests
             "Avalonia",
             "DevProjex.Avalonia",
             "MainWindow.axaml.cs");
+        var outputSource = ReadRepositoryFile(
+            "Apps",
+            "Avalonia",
+            "DevProjex.Avalonia",
+            "MainWindow.TextOutput.cs");
 
         Assert.Contains("CancelPreviewRefresh();", exportSource, StringComparison.Ordinal);
         Assert.Contains("await Task.Run(", exportSource, StringComparison.Ordinal);
@@ -194,8 +199,8 @@ public sealed class ProjectCopyExportUiContractTests
             "if (!_viewModel.CanTogglePreview || !_viewModel.IsPreviewMode)",
             windowSource,
             StringComparison.Ordinal);
-        Assert.Contains("BeginOutputPreparationStatus()", windowSource, StringComparison.Ordinal);
-        Assert.Contains("if (_viewModel.IsProjectCopyExportInProgress)", windowSource, StringComparison.Ordinal);
+        Assert.Contains("BeginOutputPreparationStatus()", outputSource, StringComparison.Ordinal);
+        Assert.Contains("if (_viewModel.IsProjectCopyExportInProgress)", outputSource, StringComparison.Ordinal);
     }
 
     [Fact]
